agents/requirements.txt
livekit-agents>=1.0.0
livekit-plugins-google>=1.0.0
livekit-plugins-silero>=1.0.0
livekit>=1.0.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
reportlab>=4.0.0
markdown>=3.5.0
